feat: improve configurability
