+++ src/config/contacts.js
@@ -1,0 +1,26 @@
+export const address = '8000 Edgewater Dr., Oakland, CA 94621, USA';
+export const contacts = [
+  {
+    type: 'text',
+    title: 'Address',
+    value: address,
+  },
+  {
+    type: 'phone',
+    title: 'Phone',
+    value: '+1 (650) 209-0-269',
+  },
+  {
+    type: 'email',
+    title: 'Email',
+    value: 'info@antrobotics.tech',
+  },
+  {
+    type: 'link',
+    title: 'LinkedIn',
+    value: 'https://www.linkedin.com/company/antrobotics',
+  },
+];
+
+export const copyright = '© ANT Robotics, 2020.';
+export const title = 'Contacts';
