+++ src/config/section-titles.js
@@ -1,6 +1,9 @@
 export const sectionTitles = {
   products: 'Wide choice of transport robots',
   benefits: 'Benefits',
+  gallery: 'Gallery',
   'cards-with-images': 'Customers benefit despite of the automation level',
   team: 'Team',
+  splash: 'Want to work with us?',
+  footer: 'Contacts',
 };
