--- src/utils/config.js
@@ -1,74 +1,0 @@
-export const benefits = [
-  {
-    image: './images/benefits-icon-1.svg',
-    title: 'Top-Notch Technology',
-    text: "Once you transform your indoor logistics, we'll keep it up-to-date.",
-  },
-  {
-    image: './images/benefits-icon-1.svg',
-    title: 'Top-Notch Technology',
-    text: "Once you transform your indoor logistics, we'll keep it up-to-date.",
-  },
-  {
-    image: './images/benefits-icon-1.svg',
-    title: 'Top-Notch Technology',
-    text: "Once you transform your indoor logistics, we'll keep it up-to-date.",
-  },
-  {
-    image: './images/benefits-icon-1.svg',
-    title: 'Top-Notch Technology',
-    text: "Once you transform your indoor logistics, we'll keep it up-to-date.",
-  },
-];
-
-export const products = [
-  {
-    image: './public/images/products-1-OTTO.svg',
-    title: 'OTTO 100 (pending)',
-    data: {
-      'Max payload: ': '220 lbs (100kg)',
-      'Size: ': '740 x 550 x 310 mm',
-      'Max speed: ': '4.3 mph',
-      'Running time: ': '4 hrs or 9 mi',
-      'Charging time: ': '1.2hrs (0-80%: 0.75 hrs)',
-    },
-  },
-  // {
-  //   "image": "./public/images/products-2-placeholder.svg",
-  //   "title": "Placeholder",
-  //   "data": {
-  //     "Max payload: ": "220 lbs (100kg)",
-  //     "Size: ": "740 x 550 x 310 mm",
-  //     "Max speed: ": "4.3 mph",
-  //     "Running time: ": "4 hrs or 9 mi",
-  //     "Charging time: ": "1.2hrs (0-80%: 0.75 hrs)"
-  //   },
-  // }
-];
-
-export const address = '8000 Edgewater Dr., Oakland, CA 94621, USA';
-export const contacts = [
-  {
-    type: 'text',
-    title: 'Address',
-    value: address,
-  },
-  {
-    type: 'phone',
-    title: 'Phone',
-    value: '+1 (650) 209-0-269',
-  },
-  {
-    type: 'email',
-    title: 'Email',
-    value: 'info@antrobotics.tech',
-  },
-  {
-    type: 'link',
-    title: 'LinkedIn',
-    value: 'https://www.linkedin.com/company/antrobotics',
-  },
-];
-
-export const copyright = '© ANT Robotics, 2020.';
-export const title = 'Contacts';
